backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
supabase>=2.0.0
openai>=1.60.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
jinja2>=3.1.0
pyyaml>=6.0.0
tenacity>=9.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
respx>=0.22.0
pytest-cov>=6.0.0
